ProjetsEtudiantsHackathon2026/Equipe8/Sujet3/requirements.txt
# Sujet 3 — Identification navale par signature radio
# Python 3.12 via uv (le système 3.9 est trop vieux pour scikit-learn récent)

# Core data
pandas>=2.2
numpy>=1.26
scipy>=1.13
pyarrow>=15            # Parquet pour les intermédiaires

# ML — algorithmes du cours d'A. Bogroff (cours 1, 3, 4)
scikit-learn>=1.5      # KMeans, StandardScaler, PCA, NearestNeighbors,
                       # EllipticEnvelope (MCD), IsolationForest,
                       # LocalOutlierFactor, OneClassSVM
umap-learn>=0.5        # UMAP (cours 3) — réduction de dim non linéaire
ruptures>=1.1          # Détection de ruptures CUSUM/BCPD (Q9)

# Géo
geopy>=2.4             # distance géodésique (Q8)
shapely>=2.0           # point-in-polygon (zones)

# Cartes & viz
folium>=0.17           # carte interactive (Q8, démo)
matplotlib>=3.8
plotly>=5.22

# Graphe de connaissance (bonus critère)
streamlit>=1.38
networkx>=3.3
pyvis>=0.3

# OSINT / API
requests>=2.32
feedparser>=6.0        # RSS (Q14)
websockets>=12.0       # AISStream.io (optionnel)
joblib>=1.4
python-dotenv>=1.0

# LLM (backend Mistral en priorité ; fallbacks dispos)
mistralai>=1.0
# anthropic>=0.39      # fallback (décommenter si clé Anthropic dispo)
# google-generativeai  # fallback Gemini (décommenter si besoin)
jinja2>=3.1            # fallback template si aucun LLM dispo

# Rapport / alerte
reportlab>=4.2         # PDF d'alerte
markdown>=3.6

# Boost de perf — supervisé (hors cours mais valorisé par la consigne :
# « force de proposition sur les méthodes »)
xgboost>=2.1
pyod>=2.0              # 40+ anomaly detection algos (ECOD, COPOD, HBOS, …)

# Backend du dashboard Next.js (API FastAPI)
fastapi>=0.115
uvicorn[standard]>=0.30

# Qualité (dev)
ruff>=0.6
pytest>=8.3
